Update and rename desafio-1-PGF5216.py to comp-mol/1d_monte-carlo.py
fixed redundancy and add comments
